src/api/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
black==25.1.0
bleach==6.2.0
cfgv==3.4.0
click==8.2.1
Deprecated==1.2.18
distlib==0.4.0
fastapi==0.116.1
filelock==3.19.1
flake8==7.3.0
h11==0.16.0
identify==2.6.14
idna==3.10
isort==6.0.1
limits==5.5.0
mccabe==0.7.0
mypy==1.18.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pre_commit==4.3.0
psutil==7.0.0
pycodestyle==2.14.0
pydantic==2.11.9
pydantic_core==2.33.2
pyflakes==3.4.0
python-dotenv==1.1.1
PyYAML==6.0.2
ruff==0.13.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.47.3
structlog==25.4.0
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
webencodings==0.5.1
wrapt==1.17.3
